Make functionality of spatial data group more robust

+++ src/main/java/io/SpatialDataGroup.java
@@ -6,6 +6,7 @@ import org.janelia.saalfeldlab.n5.N5FSWriter;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
@@ -64,13 +65,23 @@ public class SpatialDataGroup {
 		writer.setAttribute("/", "datasets", datasets.toArray());
 	}
 
-	public void addExistingDataset(String path, String datasetName) throws IOException {
-		Files.move(Paths.get(path, datasetName), Paths.get(rootPath, datasetName));
+	public void addExistingDataset(String path) throws IOException {
+		Path oldPath = Paths.get(path);
+		String datasetName = oldPath.getFileName().toString();
+
+		if (readOnly)
+			throw new SpatialDataIOException("Trying to modify a read-only spatial data group.");
+		if (datasets.contains(datasetName))
+			throw new SpatialDataIOException("Dataset '" + datasetName + "' already exists.");
+
+		Files.move(oldPath, Paths.get(rootPath, datasetName));
 		datasets.add(datasetName);
 		updateDatasetMetadata();
 	}
 
 	public void deleteDataset(String datasetName) throws IOException {
+		if (readOnly)
+			throw new SpatialDataIOException("Trying to modify a read-only spatial data group.");
 		Files.delete(Paths.get(rootPath, datasetName));
 		datasets.remove(datasetName);
 		updateDatasetMetadata();
